pyproject.toml
[project]
dependencies = [
  "fastapi[standard]",
  "sqlalchemy[asyncio]",
  "alembic",
  "asyncpg",
  "redis[hiredis]",
  "pydantic-settings",
  "python-jose[cryptography]",
  "httpx",
  "bcrypt>=5.0.0",
]
name = "sniptw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]
